# openapi.yaml
openapi: 3.1.0
info:
  title: course_service API
  version: 0.1.0
  description: |
    Контрактный API для контекста создания и доставки курсов.
    Соответствует границам Clean Architecture и DDD.
    Авторизация Bearer JWT: issuer=auth_service, audience=platform_clients, typ=access.
servers:
  - url: http://localhost:8000
...
paths:
...
  /v1/admin/courses:
    post:
      tags: [admin]
      summary: Создать курс (админ)
      operationId: adminCreateCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/admin/courses/{course_id}:
    get:
      tags: [admin]
      summary: Получить курс по ID
      operationId: adminGetCourseById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Курс найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [admin]
      summary: Обновить курс (админ)
      operationId: adminUpdateCourse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseRequest'
      responses:
        '200':
          description: Курс обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceToken:
      type: apiKey
      in: header
      name: X-Service-Token
  parameters:
    CourseId:
      name: course_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    AccessDenied:
      description: Доступ запрещен
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationError:
      description: Ошибка валидации
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
    PublishState:
      type: string
      enum: [draft, published, archived]
...
    ScheduleWindow:
      type: object
      properties:
        opens_at:
          type: string
          format: date-time
        due_at:
          type: string
          format: date-time
        enrollment_opens_at:
          type: string
          format: date-time
        enrollment_closes_at:
          type: string
          format: date-time
        starts_at:
          type: string
          format: date-time
        duration_days:
          type: integer
          minimum: 1
        access_ttl_days:
          type: integer
          minimum: 1
        timezone:
          type: string
    SeoMetadata:
      type: object
      properties:
        meta_title:
          type: string
          maxLength: 70
        meta_description:
          type: string
          maxLength: 160
        canonical_url:
          type: string
          format: uri
        robots:
          type: string
          enum: [index, noindex]
        og_image_url:
          type: string
          format: uri
    CreateCourseRequest:
      type: object
      required: [title, teacher_id, starts_at, duration_days]
      properties:
        title:
          type: string
        teacher_id:
          type: string
          format: uuid
        teacher_display_name:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
        starts_at:
          type: string
          format: date-time
        duration_days:
          type: integer
          minimum: 1
        access_ttl_days:
          type: integer
          minimum: 1
        enrollment_opens_at:
          type: string
          format: date-time
        enrollment_closes_at:
          type: string
          format: date-time
        price:
          type: number
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        language:
          type: string
        age_min:
          type: integer
          minimum: 0
        age_max:
          type: integer
          minimum: 0
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
        cover_image_url:
          type: string
          format: uri
        is_live_enabled:
          type: boolean
        live_room_template_id:
          type: string
        timezone:
          type: string
        max_students:
          type: integer
          minimum: 1
        schedule:
          $ref: '#/components/schemas/ScheduleWindow'
        seo:
          $ref: '#/components/schemas/SeoMetadata'
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: string
        teacher_id:
          type: string
          format: uuid
        teacher_display_name:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
        starts_at:
          type: string
          format: date-time
        duration_days:
          type: integer
          minimum: 1
        access_ttl_days:
          type: integer
          minimum: 1
        enrollment_opens_at:
          type: string
          format: date-time
        enrollment_closes_at:
          type: string
          format: date-time
        price:
          type: number
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        language:
          type: string
        age_min:
          type: integer
          minimum: 0
        age_max:
          type: integer
          minimum: 0
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
        cover_image_url:
          type: string
          format: uri
        is_live_enabled:
          type: boolean
        live_room_template_id:
          type: string
        timezone:
          type: string
        max_students:
          type: integer
          minimum: 1
        seo:
          $ref: '#/components/schemas/SeoMetadata'
...
    CourseResponse:
      type: object
      required:
        [course_id, title, teacher_id, starts_at, duration_days, modules_count, lessons_total, estimated_duration_hours, is_free, publish_state]
      properties:
        course_id:
          type: string
          format: uuid
        title:
          type: string
        teacher_id:
          type: string
          format: uuid
        teacher_display_name:
          type: string
          nullable: true
        slug:
          type: string
        description:
          type: string
        starts_at:
          type: string
          format: date-time
        duration_days:
          type: integer
          minimum: 1
        access_ttl_days:
          type: integer
          minimum: 1
          nullable: true
        enrollment_opens_at:
          type: string
          format: date-time
          nullable: true
        enrollment_closes_at:
          type: string
          format: date-time
          nullable: true
        price:
          type: number
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        language:
          type: string
        age_min:
          type: integer
          minimum: 0
          nullable: true
        age_max:
          type: integer
          minimum: 0
          nullable: true
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
        cover_image_url:
          type: string
          format: uri
          nullable: true
        is_live_enabled:
          type: boolean
        live_room_template_id:
          type: string
          nullable: true
        timezone:
          type: string
        max_students:
          type: integer
          minimum: 1
          nullable: true
        modules_count:
          type: integer
          minimum: 0
        lessons_total:
          type: integer
          minimum: 0
        estimated_duration_hours:
          type: integer
          minimum: 0
        is_free:
          type: boolean
        published_at:
          type: string
          format: date-time
          nullable: true
        published_by_admin_id:
          type: string
          format: uuid
          nullable: true
        archived_at:
          type: string
          format: date-time
          nullable: true
        archived_by:
          type: string
          format: uuid
          nullable: true
        publish_state:
          $ref: '#/components/schemas/PublishState'
        seo:
          $ref: '#/components/schemas/SeoMetadata'
...